Added Auto-Fit Budget Button
The program will adjust the target budgets to meet the current revenue. All of the checked budget items will have the same target funding percentage afterwards (ex. all checked programs get 90% target funding in the control panel and are 100% funded in the output panel after adjusting to the current revenue)