Values saved and reloaded when changing tax plans
If you change tax plans the slider values for slant and flat tax are saved to the tax plans list and reloaded when the plan is reloaded. NOTE: Pressing the save button on an existing plan doesn't work and resets the tax slider values. Also custom bracket adjustments aren't saved