Partially fixed bug
Saving an existing tax plan doesn't work correctly. I think it has to do with the tax rates being an observable collection and it isn't being updated properly. Right now if you set values for the slant tax, save it, for slant tax and then save it again then it will work properly the second time. In general the changes are delayed because it isn't updating properly.